+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning].
 
 - Update pre-commit hook versions
   - JuliaFormatter 1.0.58 -> 1.0.59
+- Default Indentation changed from 2 to 4 (#403)
 
 ## [0.9.1] - 2024-07-24